Replace unsafe string copy with uqm::strncpy_safe
Added uqm::strncpy_safe to safely copy strings with null-termination and buffer overflow protection. Replaced some uses of strcpy, strncpy, and similar functions across the codebase with this new utility.
Added comprehensive unit tests for strncpy_safe covering edge cases. Improved code safety, consistency, and maintainability by standardizing string copying. Also made minor improvements to enum arithmetic and variable initialization.

+++ src/uqm/planets/report.cpp
@@ -108,7 +108,7 @@ MakeReport(SOUND ReadOutSounds, uqm::CHAR_T* pStr, uqm::COUNT StrLen)
 	uqm::COUNT first_line_length = 0;
 
 	fmt::format_to_n(end_page_buf, sizeof(end_page_buf) - 1, "{}\n",
-					 GAME_STRING(SCAN_STRING_BASE + NUM_SCAN_TYPES));
+					 GAME_STRING(static_cast<int>(SCAN_STRING_BASE) + static_cast<int>(NUM_SCAN_TYPES)));
 	end_page_len = utf8StringCount(end_page_buf);
 
 	GetFrameRect(SetAbsFrameIndex(SpaceJunkFrame, 18), &r);
